// src/worker-dom-types.ts
type DOMObject = HTMLElement | SVGElement | Element | DocumentFragment | Node;
type ElementTagName = keyof HTMLElementTagNameMap | (string & {});

type WorkerDOMRef = number;
type WorkerDOMObject = DOMObject;
type WorkerDOMMethod =
  | "getRoot"
  | "create"
  | "frag"
  | "query"
  | "queryAll"
  | "childNodes"
  | "children"
  | "innerHTML"
  | "innerText"
  | "textContent"
  | "hasAttribute"
  | "getAttribute"
  | "getAttributes"
  | "setAttribute"
  | "setAttributes"
  | "removeAttribute"
  | "removeAttributes"
  | "getStyle"
  | "setStyle"
  | "append"
  | "appendChild";

interface WorkerDOMDefaultConfig {
  allowStyling: boolean;
  allowUnknown: boolean;
  allowCustomElements: boolean;
  allowedElements: Set<string>;
  forbiddenElements: Set<string>;
  forbiddenAttributes: Set<string>;
  forbiddenAttributesRgx: RegExp[];
  allowInnerHTML: boolean;
}